stats: implement invalid quartiles
When plotting quartiles without values (for example all
dives in a bin had no temperature readings), a box at 0
was shown.

To avoid this, implement the concept of invalid quartiles.
Moreover, do the conversion of bins to quartiles in the
StatsType class and chop off the begin and end of the range
if quartiles are invalid.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>