build-system/android: handle resource issues
It appears that the Kirigami shaders aren't bundled with the app. They
should be part of the plugin, but somehow they aren't. This way things
at least 'mostly work'.

We also need the icons. And to make this a bit more structured, move
those resource declarations into the Android part of the qmake file
until we know how all this works out on iOS.

The Android app is still fairly unusable with all kinds of weird font
problems and many other issues, but at least it once again starts.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>